Add module file to separate class
