+++ src/TaskClass.js
@@ -1,0 +1,9 @@
+class Task {
+  constructor(id, description, completed) {
+    this.id = id;
+    this.description = description;
+    this.completed = completed;
+  }
+}
+
+export default Task;
