+++ src/index.js
@@ -1,12 +1,12 @@
-/* eslint-disable max-classes-per-file */
 import './style.css';
 // import { update } from 'lodash';
 import Dots from './images/dots.png';
 import Enter from './images/enter.png';
-import Bin from './images/Bin.png';
+import Bin from './images/bin.png';
 import Refresh from './images/refresh.png';
 import Store from './StoreClass.js';
-import { remove } from 'lodash';
+import Task from './TaskClass.js';
+// import { remove } from 'lodash';
 // import { functionsIn, update } from 'lodash';
 
 // Get Image for refresh icon
@@ -27,19 +27,19 @@ const dots = new Image();
 dots.src = Dots;
 
 // task Class: represents a task object
-class Task {
-  constructor(id, description, completed) {
-    this.id = id;
-    this.description = description;
-    this.completed = completed;
-  }
-}
+
 // Store Class: Handles Storage
 
 // Creating new Store instance
 const store = new Store();
 
 // Update Task
+
+function getUItasks() {
+  const UiTasks = document.querySelectorAll('.element');
+  return UiTasks;
+}
+
 function updateTask(e) {
   const text = e.target.value;
   const value = e.target.id;
@@ -58,6 +58,29 @@ function updateTask(e) {
 
 const inputTasks = [];
 const buttonsDots = [];
+
+function toggleLi(e) {
+  let li;
+  const { target } = e;
+  const className = e.target.classList[0];
+  if (e.target.type === 'checkbox' || e.target.type === 'task-btn') {
+    return;
+  }
+  target.style.backgroundColor = 'yellow';
+  if (className === 'task-input') {
+    li = target.parentElement.parentElement;
+    li.style.backgroundColor = 'yellow';
+  }
+
+  if (!li) {
+    li = target;
+  }
+  const image = li.getElementsByTagName('img')[0];
+  image.setAttribute('src', Bin);
+  image.setAttribute('type', 'bin');
+  image.setAttribute('class', 'bin');
+}
+
 // UI Class : Handles UI tasks
 class UI {
   // Static so I don't have to instantiate
@@ -84,7 +107,14 @@ class UI {
     taskContent.classList.add('element');
     taskContent.setAttribute('id', `task-${task.id}`);
     const dotsContainer = taskContent.querySelector('.dots-container');
-    dotsContainer.addEventListener('click', checkTask);
+    dotsContainer.addEventListener('click', (e) => {
+      if (e.target.className === 'bin') {
+        const fullid = e.target.id;
+        const idString = fullid.split('-')[2];
+        const id = parseInt(idString, 10);
+        UI.removeTask(id);
+      }
+    });
     buttonsDots.push(dotsContainer);
   }
 
@@ -113,55 +143,17 @@ class UI {
   static clearFields() {
     document.querySelector('#list-item').value = '';
   }
-}
-function toggleLi(e) {
-  let li = undefined;
-  const target = e.target;
-  const className = e.target.classList[0];
-  if (e.target.type === 'checkbox' || e.target.type === 'task-btn') {
-    return;
-  }
-  target.style.backgroundColor = 'yellow';
-  if (className === 'task-input') {
-    li = target.parentElement.parentElement;
-    li.style.backgroundColor = 'yellow';
-  }
 
-  if (!li) {
-    li = target;
+  static removeTask(id) {
+    const currentList = store.getList();
+    const filteredList = currentList.filter((task) => task.id !== id);
+    localStorage.setItem('list', JSON.stringify(filteredList));
+    store.resetIds();
+    const tasks = getUItasks();
+    tasks.forEach((task) => task.remove());
+
+    UI.displayList();
   }
-  const image = li.getElementsByTagName('img')[0];
-  image.setAttribute('src', Bin);
-  image.setAttribute('type', 'bin');
-  image.setAttribute('class', 'bin');
-}
-
-function checkTask(e) {
-  if (e.target.className === 'bin') {
-    const fullid = e.target.id;
-    const idString = fullid.split('-')[2];
-    const id = parseInt(idString, 10);
-    removeTask(id);
-  }
-}
-
-function removeTask(id) {
-  const currentList = store.getList();
-  console.log(currentList);
-  const filteredList = currentList.filter((task) => {
-    return task.id !== id;
-  });
-  console.log(filteredList);
-  localStorage.setItem('list', JSON.stringify(filteredList));
-  const tasks = getUItasks();
-  console.log(tasks);
-  tasks.forEach((task) => task.remove());
-  UI.displayList();
-}
-
-function getUItasks() {
-  const UiTasks = document.querySelectorAll('.element');
-  return UiTasks;
 }
 
 // Update Task
